requirements/local.txt
-r base.txt

django-debug-toolbar==1.4
django-extensions
jupyter
